refactor: Reorganized project
