+++ app/pages/device_status.py
