+++ app/services/api.py
@@ -3,8 +3,9 @@
 # This program is licensed under the Apache License 2.0.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0> for full license details.
 
-import config as cfg
 from pyroclient import Client
+
+import config as cfg
 
 __all__ = ["api_client"]
